[fleet installer] Try to fix Gac Installer issue when older versions are already installed (#6715)
## Summary of changes

Fixes the issue of removing old GAC entries when there are multiple
assemblies in the GAC with the same name.

## Reason for change

https://github.com/DataDog/dd-trace-dotnet/pull/6714 highlighted an
issue with the GAC uninstall where we were _not_ removing the previous
`Datadog.Trace` entries during uninstall. This attempts to work around
that.

## Implementation details

Use another `Fusion` API to enumerate all the "Datadog.Trace"
assemblies, to get their fullnames (e.g. `Datadog.Trace,
Version=3.12.0.0, Culture=neutral, PublicKeyToken=def86d061d0d2eeb,
processorArchitecture=MSIL`).

We then run the uninstall for _all_ of the GAC assemblies we find. As
long as each assembly has been registered with a _different_ associated
filepath (i.e. the associated native loader for the gac'd file), then we
_should_ only ever remove a single file at most. The integration tests
demonstrate this works, even if it's a bit messy.

## Test coverage

Unskipped the previously skipped GAC tests, and they pass so 🤞 

## Other details

Initially I wanted to use `CreateInstallReferenceEnum()` to enumerate
the _actual_ references so that we could be sure we _only_ try to remove
the correct reference, but I couldn't get that to work (Getting an
HResult error 🤷‍♂️). The current approach appears to work though, so
maybe we should just consider that an optimisation for later? 🤷‍♂️ I
also haven't looked to see if we have any existing implementation of the
API...

Stacked on
- https://github.com/DataDog/dd-trace-dotnet/pull/6714

---------

Co-authored-by: Kevin Gosse <kevin.gosse@datadoghq.com>