add kafka topic to messaging.destination.name tag (#6758)
## Summary of changes

Sets the `messaging.destination.name` tag to the Kafka topic name for
the auto-instrumented Kafka tracer. This results in the inferred service
name of a trace to be the Kafka topic, rather than the broker.

## Reason for change

For an auto instrumented tracer, the inferred service name for kafka
requests uses the precedence defined
[here](https://docs.datadoghq.com/tracing/services/inferred_services/?tab=agentv7551#precedence-of-peer-tags).
The `peer.messaging.destination` peer tag is mapped from
`messaging.destination.name`, which is the topic name for Kafka.
This then provides better fidelity for trace propagation in the service
map, as we correlate a trace with a given topic, rather than the Kafka
broker.
This has already been done for other tracers, i.e.
[python](https://github.com/DataDog/dd-trace-py/pull/12593).

## Implementation details
Added the tag in Kafka auto-instrumentation on produce & consume side.
Consistent with other tracer implementations.

## Test coverage
* Updated integration tests
* Ran this version of the tracer on a local test app, and verified in
APM that we populate the new tag

## Other details
N/A