[Tracer] Fix Rabbitmq7 header injection overwrites user supplied basic properties [from external PR #6730] (#6753)
Based on #6730 by @johang88.

## Summary of changes
Fixes the header injection so that basic properties are preserved from
the supplied argument list in case `PopulateBasicPropertiesHeaders`
returns null which happens if the basic properties argument is writable
or if no changes had to be made to it.

## Reason for change
Fixes a bug that would remove all user supplied basic properties.

## Implementation details
Stores the basicproperties argument in the active scope and retrieves it
a null return value is encountered. Changed
`CachedBasicPropertiesHelper` to call the copy constructor that takes a
`IReadOnlyBasicProperties` as argument.

## Test coverage
Added basic properties to one of the rabbitmq7 test methods as well as
assertions.

## Other details
Fixes #6723

---------

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>
Co-authored-by: Johan Gustafsson <johan.gustafsson@bejoynd.com>