Remove instrumentation for Protobuf Schema tagging (#6792)
## Summary of changes

Attempts to remove the Protobuf Schema tagging instrumentation that was
recently added in https://github.com/DataDog/dd-trace-dotnet/pull/6647

Recommended to revert this PR when we do the fix.

## Reason for change

Appears to be throwing errors due to MissingMethodExceptions on the
DuckTypes.
We need time to investigate it and fix.
This seems to be the simplest approach as there were multiple PRs.

I don't see any code that hits this now so should be good.

Additionally, the Sample project is very flaky I attempted to remove
that code generating the proto file.

## Implementation details

- Comment out instrumented method attributes to disable instrumentations
and leave them there for less confusion (maybe?)
- Remove Proto file generation in sample which I think is the cause of
the CI build error
- Skip tests in CI.

## Test coverage

- Skipped now.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->