restore Protobuf instrumentation for schema tracking (#6826)
## Summary of changes

revert the revert of #6792
The missingMethodExceptions were most likely caused by a trimmed app
([see
discussion](https://dd.slack.com/archives/C07TBLYDQHX/p1741907211362359)).
We had some defense against this for _instrumented_ types, but not for
duck-typed types, and this is where the issue laid. This was taken care
of in #6807 for the generic case, and also after [talking with
DSM](https://dd.slack.com/archives/C069TJPTFPG/p1743500681812779), they
agree that we should try to instrument only user-defined protobufs, so I
excluded the microsoft protobufs (that were causing the issue). Thus,
this PR should be fine to merge as-is, even though it's best to merge it
after #6807 for maximum safety.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->