[ASM][ATO] User events sdk v2 methods  (#6760)
This pull request primarily focuses on the addition of
`EventTrackingSdkV2` and modifications to existing event tracking
integrations. The most important changes include the addition of the
`EventTrackingSdkV2` class, and more telemetry recording.

Why a v2? there's an RFC, but basically, so that the user provides a
user login. The user id is optional and shouldn't be the focus of
attention anymore.

### Updates to `supported_calltargets.g.json`:

* Changed the `InstrumentationTypeName` for multiple integrations to
include the `EventTrackingSdk` namespace.
[[1]](diffhunk://#diff-718a6386613bb79d4aee908359e1e0104387ddd02d2618d523fa6681b7fc175eL4487-R4487)
[[2]](diffhunk://#diff-718a6386613bb79d4aee908359e1e0104387ddd02d2618d523fa6681b7fc175eL4512-R4512)
[[3]](diffhunk://#diff-718a6386613bb79d4aee908359e1e0104387ddd02d2618d523fa6681b7fc175eL4537-R4537)
[[4]](diffhunk://#diff-718a6386613bb79d4aee908359e1e0104387ddd02d2618d523fa6681b7fc175eL4563-R4563)
[[5]](diffhunk://#diff-718a6386613bb79d4aee908359e1e0104387ddd02d2618d523fa6681b7fc175eL4587-R4587)
[[6]](diffhunk://#diff-718a6386613bb79d4aee908359e1e0104387ddd02d2618d523fa6681b7fc175eL4612-R4718)

### Addition of `EventTrackingSdkV2` class:

* Introduced the `EventTrackingSdkV2` class with methods for tracking
user login success and failure events, including overloads for different
parameter types.
[[1]](diffhunk://#diff-2a5412918fb53c028caa85eef0783559984d80dafb2ea7aed402629222fddbd7R1-R61)
[[2]](diffhunk://#diff-5d7774048584d4c3cb14215b9a31b0720765cbd6e97b6634e38255334fd08600R1-R242)

### Enhancements to telemetry recording:

* Added telemetry recording for user event metrics in
`EventTrackingSdk.cs` and `EventTrackingSdkV2.cs` to track the usage of
the SDK and specific events.
[[1]](diffhunk://#diff-41202a7c203ee4e9d646bc1725b29b03012bc12bf5c4427856c2b84d1a03a780R45-R46)
[[2]](diffhunk://#diff-41202a7c203ee4e9d646bc1725b29b03012bc12bf5c4427856c2b84d1a03a780R148-R149)
[[3]](diffhunk://#diff-41202a7c203ee4e9d646bc1725b29b03012bc12bf5c4427856c2b84d1a03a780R208-R209)

### Namespace adjustments:

* Renamed and adjusted namespaces for several integration classes to
align with the new `EventTrackingSdk` structure.

These changes collectively enhance the event tracking capabilities and
telemetry recording for user events within the `Datadog.Trace`
project.## Summary of changes

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->