Add reference to DiffPlex and use for Verify test comparisons. (#6832)
## Summary of changes

Use
[DiffPlex](https://github.com/VerifyTests/Verify.DiffPlex?tab=readme-ov-file#diff-results)
to compare verify snapshots

## Reason for change

Currently, when snapshots fail in CI, verify dumps the full diff to the
console. This makes it very hard to see what's changed, and generally
requires grabbing the full snapshots or manually diffing the contents.
Using Diffplex should make it easier to see what's changed

## Implementation details.

_Ideally_ we would just add a reference to `Verify.DiffPlex`. But as
they only support .NET FX 4.8+ and .NET 6+, we can't. So instead, add a
reference to `DiffPlex` directly, and then vendor + tweak the verify
integration.

## Test coverage

I expect something will flake in this PR, so we can see an example 😅 

## Other details