Add area filter to the pipeline (#6820)
## Summary of changes

This PR parallelizes the windows tracer and ASM tests. The goal is to
decrease the total CI running time.

This approach will also decrease the pipeline complexity by removing one
stage and nuke task.

In Linux, we have observed that benefits from this approach are not so
clear and might require a different approach, so this will be addressed
in a future/separate PR.

Basically, we are adding area traits to the integration tests assemblies
and filtering the tests based on that trait when running the integration
tests, allowing us to run them in parallel.

## Reason for change

The current CI takes too long to complete.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->