[fleet installer] Allow pushing Windows SSI OCI images to prod (#6716)
## Summary of changes

- Removes the "don't publish Windows OCI images to prod" guardrails
- Fixes the publish to prod script

## Reason for change

We want to publish prod OCI images. When testing this in a hotfix
release, I discovered there were issues with the OCI generation (we
weren't downloading all the required resources, and others were named
incorrectly). Those were fixed in the hotfix releases, so this contains
the backport of those fixes.

## Implementation details

- Delete the guardrails
- Fix the name of the fleet-installer.exe downloaded file
- Download the windows-tracer-home.zip file

## Test coverage

There's not a good or thorough way to test this without doing a release,
so that's what I did. Got it right ~first~ ~second~ third time 😉