[Test Optimization] Refactor and improvements on ImpactedTestsModule (#6836)
## Summary of changes

This pull request includes several changes to the `Datadog.Trace.Ci`
namespace to enhance the test optimization features, particularly
focusing on impacted tests detection. The most important changes involve
modifications to the `ImpactedTestsModule` class, updates to the
`TestOptimization` class, and adjustments to the `GitCommandHelper` and
`ITestOptimizationImpactedTestsDetectionFeature` interfaces.

## Reason for change

These changes collectively improve the functionality and maintainability
of the impacted tests detection feature in the Datadog CI environment.

## Implementation details

Improvements to impacted tests detection:

*
[`tracer/src/Datadog.Trace/Ci/ImpactedTestsModule.cs`](diffhunk://#diff-7bc001849f2c12b1b8887af5064d8ed71b04db32ea5d6c083bce481d03366f79L5-R221):
Refactored the `ImpactedTestsModule` class to include properties for
`BaseCommitSha`, `CurrentCommitSha`, and `ModifiedFiles`. Added methods
for creating instances and analyzing tests.
*
[`tracer/src/Datadog.Trace/Ci/TestOptimizationImpactedTestsDetectionFeature.cs`](diffhunk://#diff-48b31cb01eae55db4b261f0aa762f7ae025c637316bd6f1625096c9e21f4d0bdR19-R24):
Introduced `_environmentValues` and `_impactedTestsModule` fields, and
updated the constructor to accept `CIEnvironmentValues`.

Updates to test optimization initialization:

*
[`tracer/src/Datadog.Trace/Ci/TestOptimization.cs`](diffhunk://#diff-5b6bc242c304a5dcc672b5e713ba89f8da74b526447ea03cfc843f6758636817L272-R278):
Modified the `InitializeDefaultFeatures` method to include
`CIEnvironmentValues` as a parameter, and updated related method calls
accordingly.
[[1]](diffhunk://#diff-5b6bc242c304a5dcc672b5e713ba89f8da74b526447ea03cfc843f6758636817L272-R278)
[[2]](diffhunk://#diff-5b6bc242c304a5dcc672b5e713ba89f8da74b526447ea03cfc843f6758636817L317-R317)
[[3]](diffhunk://#diff-5b6bc242c304a5dcc672b5e713ba89f8da74b526447ea03cfc843f6758636817L553-R553)
[[4]](diffhunk://#diff-5b6bc242c304a5dcc672b5e713ba89f8da74b526447ea03cfc843f6758636817L582-R582)
[[5]](diffhunk://#diff-5b6bc242c304a5dcc672b5e713ba89f8da74b526447ea03cfc843f6758636817L600-R600)
[[6]](diffhunk://#diff-5b6bc242c304a5dcc672b5e713ba89f8da74b526447ea03cfc843f6758636817L625-R631)

Enhancements to Git command handling:

*
[`tracer/src/Datadog.Trace/Ci/GitCommandHelper.cs`](diffhunk://#diff-b7b4787860ea9ca8442f735f5f0223d6b7231dc13b96cf63a336c06c60468e37L94-R95):
Simplified the condition for checking the output of the Git command in
the `GetGitDiffFilesAndLines` method.

Interface updates:

*
[`tracer/src/Datadog.Trace/Ci/ITestOptimizationImpactedTestsDetectionFeature.cs`](diffhunk://#diff-5020ed061f4c97698c50f61d6ebc81e8422e92b9aba89930e0207fa433337fbfR14-R15):
Added a new property `ImpactedTestsAnalyzer` to the
`ITestOptimizationImpactedTestsDetectionFeature` interface.

## Test coverage

Existing test are enough

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->