Parallel Linux integration tests in the CI (#6838)
## Summary of changes

This PR is a continuation of this one:
https://github.com/DataDog/dd-trace-dotnet/pull/6820

This PR parallelizes the Linux tracer and ASM tests. The goal is to
decrease the total CI running time.

We are using the area traits of the integration tests assemblies and
filtering the tests based on that trait when running the integration
tests, allowing us to run them in parallel.

We are not including arm_64 tests since the benefits are not so clear.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->