Analyze CI crashes on Windows (#6853)
## Summary of changes

Automatically analyze crashes in integration tests on Windows, and
display the crashing stacktrace in the test output.

## Reason for change

First step to be able to skip known crashes causes by bugs in the .NET
runtime.

## Implementation details

When running a sample, we set two more environment variables:
- `DD_CRASHTRACKING_FILTERING_ENABLED=0`: tells crashtracking to process
all crashes, not just those that are caused by our instrumentation
- `DD_INTERNAL_CRASHTRACKING_OUTPUT=...`: tells crashtracking to store
the crash report in a file instead of sending it by telemetry (it's much
easier this way because we don't have to deal with the mock agent)

When a crash happens, the `ErrorHelpers.CheckForKnownSkipConditions` is
called as usual. There is now additional logic to check if a crash
report exists. If it finds one, it resolves the symbols and display the
stacktrace. In the future, we need to add logic to check for known
frames in the resolved stacktrace and throw a `SkipException` when
appropriate.

## Test coverage

I don't think we need to add a test for a test helper?

## Other details
Next steps are:
- Implement the same thing for Linux (a bit harder to fetch/process the
symbols)
 - Add the relevant rules
 - Add support for ARM64 (either very simple or very difficult)