Enable createdump crash reports in CI (#6856)
## Summary of changes

Createdump has a feature where it generate a crash report alongside the
crash dump. The crash report contains the stacktraces in json format.
This PR enables it.

## Reason for change

Coredumps on Alpine are painful to open because Windbg can't unwind
Alpine signals, so they have to be opened on an Alpine distro.
Fortunately, WSL works.
Coredumps on Alpine ARM64 are **extremely** painful because they have to
be opened on an Alpine ARM64 distro, which WSL doesn't support.

Hopefully, this change should help us identify a crash without having to
open the coredump.

## Implementation details

Just set `DOTNET_EnableCrashReport=1`. This will generate a
`*.crashreport.json` file, which we can then copy at the same time as
the dumps.