Don't merge sample build artifacts (#6851)
## Summary of changes

Don't bother to "merge" the samples into a single artifact to reduce
download size

## Reason for change

When we moved the sample build to a separate stage (to avoid re-work and
to run the build in parallel) we added a job that would aggregate all
the different TFMs into a single artifact. This was for simplicity on
the consuming side.

However, the size of this merged sample artifact was huge, such that it
takes ~6 minutes to download again in some cases (the arm64 agents are
_much_ slower to do the download). In general, only a single TFM is ever
used anyway (because we parallelize our tests by TFM too), so this is
mostly unnecessarily wasted time.

## Implementation details

- Remove the `merge` job in the `build_samples` stage
- Add a "helper" `download-samples.yml` script to make sure you download
the different samples to the correct folder (standalone goes in `bin`,
multi-version goes in `publish` currently)
- Update usages 

## Test coverage

- Ran an "all areas" test
[here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=176344&view=logs&j=66fb5b47-c134-514a-33bf-1ca485953300)
- Ran an "all TFMs" test
[here](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=176347&view=results)

## Other details

Removing the `merge` job also speeds up the `build_samples` stage which
is sometimes a blocker for testing stages, so that's an extra bonus