Refactor Flag to skip method execution (#6861)
## Summary of changes

- ~Remove the "kill switch" for the "skip method execution" filter~
- Refactor the `CallTargetState` constructors

## Reason for change

https://github.com/DataDog/dd-trace-dotnet/pull/6843 introduced a new
flag to allow bypassing the target method. This required introducing a
new parameter, `skipMethodBody`, to the `CallTargetState` constructors.
However, this was added as an optional parameter, which ultimately
resulted in overload confusion and broken NLog tests + the introduction
of a "kill switch" env variable to disable the functionality during
debugging.

The fact that the overload selection was confusing indicates to me that
we should _not_ be using optional parameters here, otherwise we're just
asking for the same problem to happen again. Similarly, having the kill
switch seems dangerous, as AFAICT there's basically no scenarios where
we would want to disable this functionality globally, as it would just
break any integration that relied on it.

## Implementation details

- Add additional overloads for the common `CallTargetState` constructors
to avoid hitting the optional parameter `skipMethodBody`. It means that
if you want to set `skipMethodBody=true`, you have to pass an additional
`state` argument too, but that seems a small price to pay for the high
level of confusion otherwise
- ~Remove the `DD_INTERNAL_SKIP_METHOD_BODY_ENABLED` kill-switch key~
After further discussion, reinstated this given the native changes carry
some risk. We'll remove this as soon as we have an integration that uses
the functionality, or potentially just in the next release either way

## Test coverage

As long as all the existing tests pass, we're ok