Decrease execution time for tests using sendHealthCheck=false (#6857)
## Summary of changes

This PR improves some tests performance by checking if a sample web page
is ready when using sendHealthCheck=false without actually calling any
endpoint.

The parameter sendHealthCheck was used to avoid calling the health
endpoint. This endpoint is used to report that a webpage is running.
Previously, skipping this call would prevent tests from knowing if the
page was active, so they would wait 30 seconds. Instead, we can check if
the port is opened and, if it is, we can assume that the sample website
is working. This way, we don't need to wait that much.

There are some affected tests. The tests running time has been decreased
from around 40/45 seconds to around 15 when tested locally.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->