[Benchmarks] Ensure high RPS for Linux throughput tests (#6869)
## Summary of changes
This PR aims to ensure high RPS for Linux throughput tests after
tweaking the ressources used for benchmarking to be able to handle it
## Reason for change
Done in order to get as close as possible to the initial setup with the
original throughput tests.
## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->