Add `Xunit.Combinatorial` for Combinatorial/Pairwise testing (#6847)
## Summary of changes

This adds
[Xunit.Combinatorial](https://github.com/AArnott/Xunit.Combinatorial) to
support both Combinatorial and Pairwise testing and support to swap
between the two when: running on a release branch, running on a
non-default branch, running locally.

## Reason for change

Our current testing strategy largely resembles a full combinatorial
expansion of test configurations. This allows us to exhaustively test
out all of our various configurations that can alter how
instrumentations behave.

However, this has led to many tests running 50+ configurations, many of
which are either redundant or _extremely_ unlikely to catch a
regression.

This change allows us to swap between a combinatorial testing
methodology (our current one) or a new pairwise testing methodology to
reduce the number of tested configurations.

Using a pairwise testing strategy significantly reduces the runtime of
many tests and will help reduce time spent in CI for PRs.

As an example, the `HttpMessageHandlerTests` will run 200 test scenarios
by default.
With pairwise it will run 20.

## Implementation details

- Added `Xunit.Combinatorial` NuGet
- Added `CombinatorialOrPairwiseDataAttribute` to swap between
Combinatorial/Pairwise
- Added `USE_FULL_TEST_CONFIG` on the build pipeline that maps to the
main/release branch
- Swapped `ADO.NET` tests to use `CombinatorialOrPairwiseDataAttribute`
- Swapped HttpMessageHandler tests to use
`CombinatorialOrPairwiseDataAttribute`

## Test coverage

- These are the tests, if they pass and look good then 👍 

## Other details

This change does run the risk of us allowing in an issue in a PR that
would've otherwise been caught. It will now be caught on _after_ the PR
is merged (as all of the test configurations will run).

If this proves to be a problem we can disable this functionality easily
globally by changing `UseFullTestConfiguration()` to always return
`true`.

This is also an _opt-in_ functionality, meaning that current tests need
to be migrated over to it.

In this PR I initially tested out _manually_ doing this, this resulted
in more fine tuning of configurations to run, but was tedious. This is
less tedious.

<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->