[Tracer] update sampling formula (#6866)
## Summary of changes

This PR changes the current formula of the sampler to have the same one
as other tracers.

A system test related to sampling disabled for dotnet will also be
enabled
[here](https://github.com/DataDog/system-tests/pull/4340/files#diff-033a5695a9b33e605ab499457e172333100b31c5fb7bf4d24834d14d76bf684cR618)
as it has been
[fixed](https://github.com/DataDog/system-tests/pull/4340/files#diff-87d5cf61d21df44e8b2276bbd6c96e943933cee8b0bd61772bad8238737ecf10R169-R174).

## Reason for change

All tracers should use the same formula in order to have a consistent
and deterministic sampling.

## Implementation details

* Removal of ` % Modulo` with `Modulo = 2^64-1` that had no impact used
as the `ulong` overflows already applies `% 2^64`
* Fast path for `0` and `1`

## Test coverage

* New tests for `0`, `1` and `0.5` rates

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->