Don't run throughput tests on PRs (#6884)
## Summary of changes

Don't run the old throughput tests on PRs

## Reason for change

- We haven't had a successful run for a week
- We're migrating to the microbenchmark platform

## Implementation details

Don't run by default on PRs (though allow manual triggering). I haven't
_completely_ deleted them yet (that's done in
https://github.com/DataDog/dd-trace-dotnet/pull/6790) because we're not
completely happy with the results from the microbenchmark platform, and
so we're still using them to gate the release

## Test coverage

This is the test. As long as the throughput tests don't run, I think
we're fine