[Dynamic Instrumentation] DEBUG-3492 Remove duration from non method snapshots (#6735)
## Summary of changes
Removes the duration property from non method snapshots

## Reason for change
We have seen customers used @duration in the expression. It seems they
wanted to know how much it took for the selected line to execute and
have a metric on that. Since @duration works only for method probes
(where it shows the execution time time in milliseconds of the method)
and in line probes it’s not useful (and I’m not sure why it’s even
there) since the line is executed similar to a breakpoint in the IDE -
before the line executes. It does not surround the method in any way -
then @duration means nothing.