Update to use `9.0.203` of the .NET SDK (#6883)
## Summary of changes

Updates to use `9.0.203` (and .NET runtime `9.0.4`) instead of `9.0.102`

## Reason for change

We've been seeing a bunch of crashes in CI unrelated to the tracer, and
hoping that this improves things

## Implementation details

- Find replace `9.0.102` -> `9.0.203`
- Update the Linux x64 VM
- Update the Windows VM (generally not required, but it's been a while)
- Update the Linux arm64 VM
- Update the GitLab build Windows Container
- Switch to the other scale sets to avoid breaking other PRs

## Test coverage

This is the test. If it all builds ok, we can merge. 

## Other details

Post merge we need to:
- [ ] Merge this https://github.com/DataDog/images/pull/7014
- [ ] Update the gitlab build to use the mirrored image above (once it's
being mirrored)
- [ ] Adjust the VM allocation for the scalesets
- [ ] Fix the macro benchmarks to install 9.0.203
- [ ] Fix the micro benchmarks to install 9.0.203
- [ ] Fix the serverless benchmarks to install 9.0.203

This _may_ affect some people's Visual Studio installation. If you've
been building with `9.0.102`, you'll need to update to the latest Visual
Studio, so that you can use the `9.0.2xx` line