[Exception Replay] Added support for top-most failing local root span + Communicating failure to capture errors (#6859)
## Summary of changes
- Error Tracking is now capturing the _top most_ errored span for a
service instead of only the root span. This change needs to be reflected
in Exception Replay for complete capturing coverage.
- Better communicating failed to capture errors. There are scenarios
where we fail to capture debug info for certain errors, it might be due
to all frames belong to 3rd party code / failed to instrument all user
frames / else. In those cases we want to communicate that clearly in the
UI to improve the UX.

## Reason for change
- Support for top most errored span.
- Improving the user experience of Exception Replay.

## Other details
Fixes DEBUG-3717

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->