[Profiler] Fix IpcServer for .NET Framework (#6903)
## Summary of changes
Do not deallocate the instance of the IpcServer used in .NET Framework

## Reason for change
Could randomly crash

## Implementation details
The IpcServer instance is never deallocated and checks are made to avoid
using it when/after the CLR shutdown

## Test coverage
N/A

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->