Add support for AWS SDK v4.0 (#6929)
## Summary of changes

Adds support for AWS SDK 4.0, [released on 28th
Apr](https://aws.amazon.com/blogs/developer/general-availability-of-aws-sdk-for-net-v4-0/)

## Reason for change

We want to support the latest versions of the SDK

## Implementation details

- Ran a "baseline" bump of package versions for simplicity, and then
later removed the commit from the stack.
- Bumped the tested versions of AWS.
- Bumped the instrumented versions of AWS.
- Added scrubbing for http-client-handler vs socket-handler.
- This is an implementation detail difference in v4 which doesn't really
matter in the snapshots so just scrubbed it where required
- Added additional instrumentation to grab the "real" URL that the HTTP
Client sends.
- The value added by the existing "core pipeline" instrumentation
contained the wrong value for the S3 tests, it contained a value
`/{Key+}` instead of the expected `/sample.txt`.
- This looks like a core change to the library, to allow "parameterized"
URLs.
- Rather than trying to reproduce the logic it uses in the existing
endpoint, the additional instrumentation grabs the "real" URL that's
used, and uses that preferentially
- I haven't checked to see if we could backport this to earlier SDK
versions too, but seems like low value to do so with risks of breakage

## Test coverage

Ran an [all-TFMs
test](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=177194&view=results)
to make sure everything looks ok, and looks fine in general

## Other details

Stacked on 
- https://github.com/DataDog/dd-trace-dotnet/pull/6922

because the minimum .NET Framework version was bumped to `net472`