rewrite GenerateSpanDocumentation using roslyn to parse the files (#6937)
## Summary of changes

Removed the artisan parsing code that was brittle and subject to failure
if formatting was not perfect, and replaced it with code using roslyn to
properly parse the C# code.

## Reason for change

The existing code required expression bodies, but the editorconfig
suggest using block bodies everywhere, so changing this file was
challenging with "reformat on save" turned on.

## Implementation details

used Roslyn. It's nice.

## Test coverage

No tests for this afaik. I regenerated the markdown using the old code,
and than with the new code, and there was no change. Also manually
tested that having an expression body now works.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->